+++ src/hv_anndata/plotting/__init__.py
@@ -1,0 +1,24 @@
+"""HoloViz plotting using AnnData as input."""
+
+from __future__ import annotations
+
+from .clustermap import ClusterMap, ClusterMapConfig, create_clustermap_plot
+from .dotmap import Dotmap, DotmapConfig
+from .labeller import labeller
+from .manifoldmap import (
+    ManifoldMap,
+    ManifoldMapConfig,
+    create_manifoldmap_plot,
+)
+
+__all__ = [
+    "ClusterMap",
+    "ClusterMapConfig",
+    "Dotmap",
+    "DotmapConfig",
+    "ManifoldMap",
+    "ManifoldMapConfig",
+    "create_clustermap_plot",
+    "create_manifoldmap_plot",
+    "labeller",
+]
