+++ src/hv_anndata/plotting/manifoldmap.py
@@ -17,11 +17,12 @@ import panel as pn
 import panel_material_ui as pmui
 import param
 from bokeh.models.tools import BoxSelectTool, LassoSelectTool
-from holoviews.operation import Operation
 from holoviews.selection import link_selections
 from panel.reactive import hold
 
-from .interface import ACCESSOR as A
+from hv_anndata import ACCESSOR as A
+
+from .labeller import labeller
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
@@ -45,61 +46,12 @@ DEFAULT_CAT_CMAP = cc.b_glasbey_category10
 DEFAULT_CONT_CMAP = "viridis"
 
 
-def _is_categorical(arr: np.ndarr) -> bool:
+def _is_categorical(arr: np.ndarray) -> bool:
     return (
         arr.dtype.name in ["category", "categorical", "bool"]
         or np.issubdtype(arr.dtype, np.object_)
         or np.issubdtype(arr.dtype, np.str_)
     )
-
-
-class labeller(Operation):  # noqa: N801
-    """Add a Label element centered over categorical points."""
-
-    column = param.String()
-
-    max_labels = param.Integer(10)
-
-    min_count = param.Integer(default=1)
-
-    streams = param.List([hv.streams.RangeXY])
-
-    x_range = param.Tuple(
-        default=None,
-        length=2,
-        doc="""
-       The x_range as a tuple of min and max x-value. Auto-ranges
-       if set to None.""",
-    )
-
-    y_range = param.Tuple(
-        default=None,
-        length=2,
-        doc="""
-       The x_range as a tuple of min and max x-value. Auto-ranges
-       if set to None.""",
-    )
-
-    def _process(self, el: hv.Dataset, key=None) -> hv.Labels:  # noqa: ARG002, ANN001
-        if self.p.x_range and self.p.y_range:
-            el = el[slice(*self.p.x_range), slice(*self.p.y_range)]
-
-        df = el.dframe()
-        xd, yd, cd = el.dimensions()[:3]
-        col = self.p.column or cd.name
-        result = (
-            df.groupby(col)
-            .agg(
-                count=(col, "size"),  # count of rows per group
-                x=(xd.name, "mean"),
-                y=(yd.name, "mean"),
-            )
-            .query(f"count > {self.p.min_count}")
-            .sort_values("count", ascending=False)
-            .iloc[: self.p.max_labels]
-            .reset_index()
-        )
-        return hv.Labels(result, ["x", "y"], col)
 
 
 class ManifoldMapConfig(TypedDict, total=False):
