--- tests/test_basic.py
@@ -1,7 +1,0 @@
-"""Basic tests."""
-
-from __future__ import annotations
-
-
-def test_import() -> None:
-    import hv_anndata  # noqa: F401, PLC0415
